Add configurable option for logging file path
Adds a new option in the argus.conf, for specifying
the file path where argus should write the logging to.
The 'argus_log_file' option can be found in the argus section.
A new module has been added for logging, in order to avoid
import errors in multiple modules and also to make it easier
to manage to logging module.
