+++ argus/log.py
@@ -1,0 +1,47 @@
+# Copyright 2016 Cloudbase Solutions Srl
+# All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+import logging
+
+from argus import config as argus_config
+
+CONFIG = argus_config.CONFIG
+
+
+DEFAULT_FORMAT = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+
+
+def get_logger(name="argus",
+               format_string=DEFAULT_FORMAT,
+               logging_file=CONFIG.argus.argus_log_file):
+    """Obtain a new logger object.
+
+    The `name` parameter will be the name of the logger and `format_string`
+    will be the format it will use for logging. `logging_file` is a file
+    where the messages will be written.
+    """
+    logger = logging.getLogger(name)
+    formatter = logging.Formatter(format_string)
+
+    if not logger.handlers:
+        # If the logger wasn't obtained another time,
+        # then it shouldn't have any loggers
+
+        if logging_file:
+            file_handler = logging.FileHandler(logging_file, delay=True)
+            file_handler.setFormatter(formatter)
+            logger.addHandler(file_handler)
+
+    logger.setLevel(logging.DEBUG)
+    return logger
