+++ argus/recipes/cloud/base.py
@@ -20,14 +20,13 @@ import abc
 import six
 
 from argus import config as argus_config
+from argus import log as argus_log
 from argus.recipes import base
-from argus import util
-
 
 __all__ = ('BaseCloudbaseinitRecipe', )
 
 CONFIG = argus_config.CONFIG
-LOG = util.get_logger()
+LOG = argus_log.get_logger()
 
 
 @six.add_metaclass(abc.ABCMeta)
